update DOCKER_REPO_USER env variable
our Jenkinsfile differs somewhat from
https://github.com/kingdonb/Jenkinsfile/blob/main/Jenkinsfile

One of the expectations that differs here is we do not have any Docker
Hub credentials at build time. This stage is separate so you can imagine
that you run tests in between building the image and pushing it to the
Docker registry.

(We do not run any test in this repo, the images are just promoted
according to the shell scripts, which is also prone to defer to the
Makefile about how builds should ran.)